+++ R/MAPAssignActivity.R
@@ -92,7 +92,8 @@ MAPAssignActivity <-
            use.sparse.assign          = FALSE,
            drop.low.mut.samples       = TRUE,
            use.sig.presence.test      = FALSE,
-           sig.pres.test.nbinom.size  = NULL) {
+           sig.pres.test.nbinom.size  = NULL,
+           sig.pres.test.p.thresh     = 0.05) {
     if (drop.low.mut.samples) {
       spectra <- DropLowMutationSamples(spectra)
     } else {
@@ -119,7 +120,8 @@ MAPAssignActivity <-
         use.sparse.assign           = use.sparse.assign,
         drop.low.mut.samples        = drop.low.mut.samples, 
         use.sig.presence.test       = use.sig.presence.test,
-        sig.pres.test.nbinom.size   = sig.pres.test.nbinom.size)
+        sig.pres.test.nbinom.size   = sig.pres.test.nbinom.size,
+        sig.pres.test.p.thresh      = sig.pres.test.p.thresh)
 
       return(retval1)
     }