+++ R/RunMAPOneSample.R
@@ -26,7 +26,8 @@ RunMAPOnOneSample <-
            use.sparse.assign           = FALSE,
            drop.low.mut.samples        = TRUE,
            use.sig.presence.test       = FALSE,
-           sig.pres.test.nbinom.size   = 69) {
+           sig.pres.test.nbinom.size   = NULL,
+           sig.pres.test.p.thresh      = 0.05) {
     
     if (drop.low.mut.samples) {
       spect <- DropLowMutationSamples(spect)
@@ -53,7 +54,8 @@ RunMAPOnOneSample <-
       use.sparse.assign          = use.sparse.assign,
       drop.low.mut.samples       = drop.low.mut.samples, 
       use.sig.presence.test      = use.sig.presence.test,
-      sig.pres.test.nbinom.size  = sig.pres.test.nbinom.size)
+      sig.pres.test.nbinom.size  = sig.pres.test.nbinom.size,
+      sig.pres.test.p.thresh     = sig.pres.test.p.thresh)
     
     if (!is.null(retval$error.messages)) {
       return(retval)